requirements.txt
requests
Pillow
sixel
